[REFACTOR] Error 코드 및 조회 API 수정

+++ controller/board/deleteMyposts.js
@@ -6,16 +6,17 @@ const deleteMyposts = async (req, res, next) => {
   try {
     const postIdList = req.body.post_id_list;
 
-    for (let i = 0; i < postIdList.length; i++) {
-      const Post = await models.Post.destroy({ where: { id: postIdList[i] } });
-      if (!Post) {
-        res.status(400).send('400 Bad Request');
-      }
+    if (!postIdList) {
+      return res.status(400).send('EMPTY_POST_ID_LIST');
     }
 
-    res.status(200).send('200 OK');
+    for (let i = 0; i < postIdList.length; i++) {
+      await models.Post.destroy({ where: { id: postIdList[i] } });
+    }
+
+    return res.status(200).send('DELETE_MY_POST_SUCCESS');
   } catch (err) {
-    next(err);
+    return res.status(500).send('DELETE_MY_POST_FAILURE');
   }
 };
